requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
spacy
networkx
fastapi
uvicorn
datasets